moved Options class and tool error messages resource to impl package.
